move function to Centrifuge entity
