+++ src/index.ts
@@ -68,4 +68,3 @@ export type { Callback } from './utils/scriptHash.js'
 export { MAGIC_VARIABLE_KEYS, resolveMagicVariables, resolveVariableLabel } from './utils/variables.js'
 export type { MagicVariableContext, MagicVariableKey } from './utils/variables.js'
 export { manifestToDefinition } from './utils/workflow.js'
-export { fetchWorkflowManifest } from './utils/workflowManifest.js'
